Rename hello_world folder to hello-world
