+++ packages/open-truss/src/hello-world/index.ts
