+++ packages/open-truss/src/index.ts
@@ -1,3 +1,3 @@
-export { default as helloWorld } from './hello_world'
+export { default as helloWorld } from './hello-world'
 export * from './utils/yaml'
 export * from './utils/format'
